Do partial operations on views of dynamic_bitset
Create partial views of dynamic_bitset. Use the partial view as the base
object for the partial operations.
Eliminate the function that operates on a complete vector, and consider
it as a special case of the "general" case

A const_dynamic_bitset_span enables the creation of a const
dynamic_bitset_span when the base dynamic_bitset is const